feat(client): details page changes
