+++ libs/ui/src/next/popover.tsx
@@ -16,31 +16,24 @@ function PopoverTrigger({
 
 const PopoverContent = React.forwardRef<
 	React.ElementRef<typeof PopoverPrimitive.Content>,
-	React.ComponentProps<typeof PopoverPrimitive.Content> & {
-		container?: HTMLElement | null;
-	}
->(
-	(
-		{ className, align = "center", sideOffset = 4, container, ...props },
-		ref,
-	) => {
-		return (
-			<PopoverPrimitive.Portal container={container}>
-				<PopoverPrimitive.Content
-					ref={ref}
-					data-slot="popover-content"
-					align={align}
-					sideOffset={sideOffset}
-					className={cn(
-						"data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 z-50 w-72 origin-(--radix-popover-content-transform-origin) rounded-md border bg-popover p-4 text-popover-foreground shadow-md outline-hidden data-[state=closed]:animate-out data-[state=open]:animate-in",
-						className,
-					)}
-					{...props}
-				/>
-			</PopoverPrimitive.Portal>
-		);
-	},
-);
+	React.ComponentProps<typeof PopoverPrimitive.Content>
+>(({ className, align = "center", sideOffset = 4, ...props }, ref) => {
+	return (
+		<PopoverPrimitive.Portal>
+			<PopoverPrimitive.Content
+				ref={ref}
+				data-slot="popover-content"
+				align={align}
+				sideOffset={sideOffset}
+				className={cn(
+					"data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 z-50 w-72 origin-(--radix-popover-content-transform-origin) rounded-md border bg-popover p-4 text-popover-foreground shadow-md outline-hidden data-[state=closed]:animate-out data-[state=open]:animate-in",
+					className,
+				)}
+				{...props}
+			/>
+		</PopoverPrimitive.Portal>
+	);
+});
 
 PopoverContent.displayName = "PopoverContent";
