+++ packages/client/src/components/settings/index.ts
@@ -1,6 +1,6 @@
-export * from "./EngineModelTestSidebar";
-export * from "./EngineQASidebar";
-export * from "./FileTable";
+export * from "./engine-model-test-sidebar";
+export * from "./engine-qa-sidebar";
+export * from "./file-table";
 export * from "./members-table";
 export * from "./pending-members-table";
 export * from "./settings.types";
