+++ packages/client/src/components/teams/index.ts
@@ -1,4 +1,4 @@
-export * from "./AddTeamModal";
+export * from "./add-team-modal";
 export * from "./TeamEnginesTable";
 export * from "./TeamMembersTable";
 export * from "./TeamProjectsTable";
