--- packages/client/src/pages/engine/EngineFilePage.tsx
@@ -1,42 +1,0 @@
-import { styled, Table, Typography } from "@semoss/ui";
-import { FileTable } from "@/components/settings";
-import { useEngine } from "@/hooks";
-
-const StyledContainer = styled("div")(({ theme }) => ({
-	width: "100%",
-	display: "flex",
-	alignSelf: "stretch",
-	flexDirection: "column",
-	alignItems: "flex-start",
-	gap: theme.spacing(3),
-}));
-
-const StyledTableContainer = styled(Table.Container)({
-	borderRadius: "12px",
-	// background: #FFF;
-	/* Devias Drop Shadow */
-	boxShadow: "0px 5px 22px 0px rgba(0, 0, 0, 0.06)",
-});
-
-const StyledTopDiv = styled("div")(() => ({
-	display: "flex",
-	width: "100%",
-	justifyContent: "space-between",
-}));
-
-export const EngineFilePage = () => {
-	//Grabbing Engine Id for document creation
-	const { active } = useEngine();
-
-	return (
-		<StyledContainer>
-			<StyledTopDiv>
-				<Typography variant={"h6"}>File Explorer</Typography>
-			</StyledTopDiv>
-
-			<StyledTableContainer>
-				<FileTable id={active.id} />
-			</StyledTableContainer>
-		</StyledContainer>
-	);
-};
